% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_flow.R
\name{app_flow}
\alias{app_flow}
\title{Flow}
\usage{
app_flow(tomic)
}
\arguments{
\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}
}
\value{
A \code{shiny} app
}
\description{
Using \code{shiny} comb through datasets by iterating between
  plotting steps, and lassoing steps to select points of interest.
}
\examples{

if (interactive()) {
  # library(reactlog)
  # reactlog_enable()
  app_flow(brauer_2008_triple)
  # shiny::reactlogShow()
}

}
