% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_heatmap.R
\name{app_heatmap}
\alias{app_heatmap}
\title{Interactive Heatmap}
\usage{
app_heatmap(tomic)
}
\arguments{
\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}
}
\value{
A \code{shiny} app
}
\description{
Generate a \code{shiny} interactive heatmap that allows for on demand
  filtering, ordering and faceting by variables of interest.
}
\examples{

if (interactive()) {
  app_heatmap(brauer_2008_tidy)
}
}
