% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_scatter.R
\name{app_scatter}
\alias{app_scatter}
\title{Scatter Plot}
\usage{
app_scatter(tidy_omic)
}
\arguments{
\item{tidy_omic}{a tidy_omic object constructed from
\code{\link{create_tidy_omic}}}
}
\value{
A \code{shiny} app
}
\description{
Generate a Shiny interactive scatter plot which allows visualization of
features, measurements, and samples (with principal components added).
}
\examples{
if (interactive()) {
  app_scatter(brauer_2008_tidy)
}

}
