% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutates.R
\name{center_tomic}
\alias{center_tomic}
\title{Center T* Omic}
\usage{
center_tomic(tomic, measurement_vars = "all")
}
\arguments{
\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}

\item{measurement_vars}{measurement variables to center}
}
\value{
An object of the same class as \code{tomic_to}
}
\description{
Center each measurement by subtracting the mean.
}
\examples{
center_tomic(brauer_2008_tidy)
}
