% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_classes.R
\name{check_tidy_omic}
\alias{check_tidy_omic}
\title{Check Tidy Omic}
\usage{
check_tidy_omic(tidy_omic, fast_check = TRUE)
}
\arguments{
\item{tidy_omic}{an object of class tidy_omic produced by
\code{\link{create_tidy_omic}}}

\item{fast_check}{if TRUE then skip some checks which are slow and that are
generally only needed when a \code{tomic} object is first created.}
}
\value{
0 invisibly
}
\description{
Check a tidy omic dataset for consistency between the data and design and
  validate that the dataset follows the \code{tidy_omic}/\code{tomic}
  specification.
}
