% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_classes.R
\name{check_tomic}
\alias{check_tomic}
\title{Check T*Omic}
\usage{
check_tomic(tomic, fast_check = TRUE)
}
\arguments{
\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}

\item{fast_check}{if TRUE then skip some checks which are slow and that are
generally only needed when a \code{tomic} object is first created.}
}
\value{
0 invisibly
}
\description{
Check a tidy or triple 'omic object for common pathologies, such as a
  mismatch between data and schema and non-uniqueness of primary keys.
}
\examples{
check_tomic(brauer_2008_triple)
}
