% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_classes.R
\name{check_triple_omic}
\alias{check_triple_omic}
\title{Check Triple Omic}
\usage{
check_triple_omic(triple_omic, fast_check = TRUE)
}
\arguments{
\item{triple_omic}{an object of class triple_omic produced by
\code{\link{create_triple_omic}}}

\item{fast_check}{if TRUE then skip some checks which are slow and that are
generally only needed when a \code{tomic} object is first created.}
}
\value{
0 invisibly
}
\description{
Check a triple omic dataset for consistency between the data and design and
  validate that the dataset follows the \code{triple_omic}/\code{tomic}
  specification.
}
