% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_heatmap.R
\name{downsample_heatmap}
\alias{downsample_heatmap}
\title{Downsample Heatmap}
\usage{
downsample_heatmap(tidy_data, value_var, max_display_features = 1000)
}
\arguments{
\item{tidy_data}{The data datable from a \code{tidy_omic} object containing
ordered feature and sample primary keys defined by ordered_featureId
and ordered_sampleId.}

\item{value_var}{which variable in "measurements" to use for quantification.}

\item{max_display_features}{aggregate and downsample distinct feature to
this number to speed to up heatmap rendering.}
}
\value{
tidy_data with rows collapsed if the number of distinct features is
  greater than \code{max_display_features}
}
\description{
Combine rows to speed up rendering of large heatmaps
}
\examples{
library(dplyr)

brauer_2008_tidy$data \%>\%
  dplyr::mutate(
  ordered_featureId = factor(name, levels = unique(name)),
  ordered_sampleId = factor(sample, levels = unique(sample))
  ) \%>\%
  downsample_heatmap(value_var = "expression", 100)
}
