% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_tomic_as_tidy}
\alias{export_tomic_as_tidy}
\title{Export T*Omic in Tidy Format}
\usage{
export_tomic_as_tidy(tomic, dir_path, name_preamble, verbose = TRUE)
}
\arguments{
\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}

\item{dir_path}{path to save outputs}

\item{name_preamble}{start of output file name}

\item{verbose}{extra reporting messages}
}
\value{
Export one table which is one row per peak, which includes
  all feature and sample attributes.
}
\description{
Export a data table including all fields from features, samples and
  measurements.
}
\examples{

if (interactive()) {
  export_tomic_as_tidy(brauer_2008_triple, "/tmp", "brauer")
}
}
