% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter.R
\name{filterInput}
\alias{filterInput}
\title{Filter Input}
\usage{
filterInput(id, filter_table)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{filter_table}{table to filter}
}
\value{
A \code{shiny} UI
}
\description{
UI components for the filter module.
}
