% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter.R
\name{filterServer}
\alias{filterServer}
\title{Filter Server}
\usage{
filterServer(id, tidy_omic, filter_table)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{tidy_omic}{an object of class tidy_omic produced by
\code{\link{create_tidy_omic}}}

\item{filter_table}{table to filter}
}
\description{
Filter Server
}
