% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{get_design_tbl}
\alias{get_design_tbl}
\title{Get Design Table}
\usage{
get_design_tbl(tomic)
}
\arguments{
\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}
}
\value{
a tibble reflecting the \code{tomic} object's design.
}
\description{
Get a tabular summary of all variables.
}
\examples{
get_design_tbl(brauer_2008_triple)
}
