% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ggbiv.R
\name{ggBivOutput}
\alias{ggBivOutput}
\title{ggBivariate Output}
\usage{
ggBivOutput(id, return_brushed_points = FALSE)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{return_brushed_points}{Return values selected on the plot}
}
\value{
A \code{shiny} UI
}
\description{
UI components for the ggBivariate module.
}
