% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ggplot.R
\name{ggplotOutput}
\alias{ggplotOutput}
\title{ggplot Output}
\usage{
ggplotOutput(
  id,
  default_data_type = "samples",
  default_plot_type = "univariate"
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{default_data_type}{Default data type selection}

\item{default_plot_type}{Default plot type selection}
}
\value{
A \code{shiny} UI
}
\description{
UI components for the ggplot module.
}
