% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ggplot.R
\name{ggplotServer}
\alias{ggplotServer}
\title{ggplot Server}
\usage{
ggplotServer(id, tomic, return_brushed_points = FALSE)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}

\item{return_brushed_points}{Return values selected on the plot}
}
\value{
a \code{tibble} of selected observations if
  \code{return_brushed_points} is TRUE. Otherwise, returns NULL.
}
\description{
Server components for the ggplot module.
}
