% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_gguniv.R
\name{ggUnivOutput}
\alias{ggUnivOutput}
\title{ggUnivariate Output}
\usage{
ggUnivOutput(id, return_brushed_points = FALSE)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{return_brushed_points}{Return values selected on the plot}
}
\value{
A \code{shiny} UI
}
\description{
UI components for the ggUnivariate module.
}
