% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_gguniv.R
\name{ggUnivServer}
\alias{ggUnivServer}
\title{ggUnivariate Server}
\usage{
ggUnivServer(id, tomic, plot_table, return_brushed_points = FALSE)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}

\item{plot_table}{table containing the data to be plotted}

\item{return_brushed_points}{Return values selected on the plot}
}
\value{
a tomic_table if return_brushed_points is TRUE, and 0 otherwise.
}
\description{
Server components for the ggUnivariate module
}
