% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{impute_missing_values}
\alias{impute_missing_values}
\title{Impute Missing Values}
\usage{
impute_missing_values(
  tomic,
  impute_var_name = "imputed",
  value_var = NULL,
  ...
)
}
\arguments{
\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}

\item{impute_var_name}{variable to create for imputed measurements}

\item{value_var}{An abundance value to use with \code{hclust}}

\item{...}{additional arguments to pass to \link[impute]{impute.knn}}
}
\value{
A \code{tomic} object with imputed measurements.
}
\description{
Impute missing values using K-nearest neighbors imputation
}
\examples{
impute_missing_values(brauer_2008_triple)

}
