% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_classes.R
\name{infer_tomic_table_type}
\alias{infer_tomic_table_type}
\title{Infer Tomic Table Type}
\usage{
infer_tomic_table_type(tomic, tomic_table)
}
\arguments{
\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}

\item{tomic_table}{A table taken from a tidy (i.e., augmented measurements)
or triple omic dataset}
}
\value{
features, samples or measurements
}
\description{
From a tomic_table, choose whether it reflects features, samples or
  measurements
}
