% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_lasso.R
\name{lassoInput}
\alias{lassoInput}
\title{Lasso Input}
\usage{
lassoInput(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
A \code{shiny} UI
}
\description{
UI components for the lasso module.
}
