% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_organize.R
\name{organizeServer}
\alias{organizeServer}
\title{Organize Servers}
\usage{
organizeServer(id, tidy_omic, feature_vars, sample_vars, value_var)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{tidy_omic}{an object of class tidy_omic produced by
\code{\link{create_tidy_omic}}}

\item{feature_vars}{variables available for arranging features}

\item{sample_vars}{variables available for arrange samples}

\item{value_var}{An abundance value to use with \code{hclust}}
}
\value{
A \code{tomic} with sorted features and/or samples.
}
\description{
Server components for the organize input module.
}
