% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{plot_missing_values}
\alias{plot_missing_values}
\title{Plot Missing Values}
\usage{
plot_missing_values(tomic, value_var = NULL)
}
\arguments{
\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}

\item{value_var}{the measurement variable to check for missingness (NA or no entry)}
}
\value{
a ggplot2 grob
}
\description{
Create a simple plot of missing values.
}
\examples{
plot_missing_values(brauer_2008_triple)
}
