% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plotsaver.R
\name{plotsaverInput}
\alias{plotsaverInput}
\title{Plot Saver Input}
\usage{
plotsaverInput(id, ui_format = "tall")
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{ui_format}{Set UI appearance
\describe{
  \item{tall}{stack all UI elements}
  \item{wide}{UI elements are side-by-side}
}}
}
\value{
a \code{shiny} UI
}
\description{
UI components for the plot saver module.
}
