% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{prepare_example_datasets}
\alias{prepare_example_datasets}
\title{Format example datasets and add them to the package}
\usage{
prepare_example_datasets()
}
\description{
Format example datasets and add them to the package
}
