% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{reconcile_triple_omic}
\alias{reconcile_triple_omic}
\title{Reconcile Triple Omic}
\usage{
reconcile_triple_omic(triple_omic)
}
\arguments{
\item{triple_omic}{an object of class triple_omic produced by
\code{\link{create_triple_omic}}}
}
\value{
a triple_omic object
}
\description{
If some samples, feature or measurements have been dropped; update other
  tables.
}
