% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romic-package.R
\docType{package}
\name{romic-package}
\alias{romic}
\alias{romic-package}
\title{romic: R for High-Dimensional Omic Data}
\description{
'romic' represents high-dimensional data as tables of features,
  samples and measurements, and a design list for tracking the meaning of
  individual variables. Using this format, filtering, normalization, and
  other transformations of a dataset can be carried out in a flexible
  manner. 'romic' takes advantage of these transformations to create
  interactive shiny apps for exploratory data analysis such as an
  interactive heatmap.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shackett.org/romic-package/}
  \item \url{https://github.com/shackett/romic}
  \item Report bugs at \url{https://github.com/shackett/romic/issues}
}

}
\author{
\strong{Maintainer}: Sean Hackett \email{seanrhackett@gmail.com} (\href{https://orcid.org/0000-0002-9553-4341}{ORCID})

Other contributors:
\itemize{
  \item Calico Life Sciences LLC [copyright holder, funder]
}

}
\keyword{internal}
