% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter.R
\name{shiny_filter_test}
\alias{shiny_filter_test}
\title{Shiny Filter Test}
\usage{
shiny_filter_test(tidy_omic, filter_table = "features")
}
\arguments{
\item{tidy_omic}{an object of class tidy_omic produced by
\code{\link{create_tidy_omic}}}

\item{filter_table}{table to filter}
}
\value{
A \code{shiny} app
}
\description{
Tests the shiny filter module as a stand-alone application.
}
\examples{
if (interactive()) {
  shiny_filter_test(brauer_2008_tidy)
}
}
