% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ggbiv.R
\name{shiny_ggbiv_test}
\alias{shiny_ggbiv_test}
\title{Shiny ggBivariate Test}
\usage{
shiny_ggbiv_test(tomic, plot_table = "samples")
}
\arguments{
\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}

\item{plot_table}{table containing the data to be plotted}
}
\value{
a \code{shiny} app
}
\description{
Test the shiny ggBivariate module as a stand-alone application.
}
\examples{

if (interactive()) {
  shiny_ggbiv_test(
    add_pcs(brauer_2008_triple, npcs = 5),
    plot_table = "samples"
  )
  shiny_ggbiv_test(
    brauer_2008_triple,
    plot_table = "measurements"
  )
}
}
