% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ggplot.R
\name{shiny_ggplot_test}
\alias{shiny_ggplot_test}
\title{Shiny ggplot Test}
\usage{
shiny_ggplot_test(tomic)
}
\arguments{
\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}
}
\value{
A \code{shiny} app
}
\description{
Test the shiny ggplot module as a stand-alone application.
}
\examples{

if (interactive()) {
  shiny_ggplot_test(add_pcs(brauer_2008_triple, npcs = 5))
  shiny_ggplot_test(brauer_2008_triple)
}
}
