% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_gguniv.R
\name{shiny_gguniv_test}
\alias{shiny_gguniv_test}
\title{Shiny ggUnivariate Test}
\usage{
shiny_gguniv_test(tomic, plot_table = "samples")
}
\arguments{
\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}

\item{plot_table}{table containing the data to be plotted}
}
\value{
A \code{shiny} app
}
\description{
Test the shiny ggUnivariate module as a stand-alone application.
}
\examples{

if (interactive()) {
  shiny_gguniv_test(
    add_pcs(brauer_2008_triple, npcs = 5),
    plot_table = "samples"
  )
  shiny_gguniv_test(brauer_2008_triple, plot_table = "measurements")
  shiny_gguniv_test(brauer_2008_triple, plot_table = "features")
}
}
