% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_organize.R
\name{shiny_organize_test}
\alias{shiny_organize_test}
\title{Shiny Organize Test}
\usage{
shiny_organize_test(tidy_omic, feature_vars, sample_vars, value_var)
}
\arguments{
\item{tidy_omic}{an object of class tidy_omic produced by
\code{\link{create_tidy_omic}}}

\item{feature_vars}{variables available for arranging features}

\item{sample_vars}{variables available for arrange samples}

\item{value_var}{An abundance value to use with \code{hclust}}
}
\value{
a \code{shiny} app
}
\description{
Tests the shiny organization module as stand-alone application.
}
\examples{

if (interactive()) {
  shiny_organize_test(
    brauer_2008_tidy,
    feature_vars = c("BP", "MF"),
    sample_vars = c("sample", "nutrient", "DR"),
    value_var = "expression"
  )
}
}
