% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_sort.R
\name{shiny_sort_test}
\alias{shiny_sort_test}
\title{Shiny Sort Test}
\usage{
shiny_sort_test(triple_omic, valid_sort_vars, value_var)
}
\arguments{
\item{triple_omic}{an object of class triple_omic produced by
\code{\link{create_triple_omic}}}

\item{valid_sort_vars}{variables available for categorical arranging}

\item{value_var}{An abundance value to use with \code{hclust}}
}
\value{
a \code{shiny} app
}
\description{
Test the shiny sorting module as a stand-alone app.
}
\examples{

if (interactive()) {
  shiny_sort_test(brauer_2008_triple,
    valid_sort_vars = c("sample", "nutrient", "DR"),
    value_var = "expression"
  )
}
}
