% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutates.R
\name{sort_triple_arrange}
\alias{sort_triple_arrange}
\title{Sort Triple Arrange}
\usage{
sort_triple_arrange(triple_omic, sort_table, sort_variables)
}
\arguments{
\item{triple_omic}{an object of class triple_omic produced by
\code{\link{create_triple_omic}}}

\item{sort_table}{samples or features}

\item{sort_variables}{A set of attributes in sort_table to sort with in
\code{arrange}.}
}
\value{
A \code{triple_omic} with sorted features or samples.
}
\description{
Sort a \code{triple_omic} object based on the values of one or more
  variables.
}
