% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutates.R
\name{sort_triple_hclust}
\alias{sort_triple_hclust}
\title{Sort Triple Hclust}
\usage{
sort_triple_hclust(triple_omic, sort_table, value_var)
}
\arguments{
\item{triple_omic}{an object of class triple_omic produced by
\code{\link{create_triple_omic}}}

\item{sort_table}{samples or features}

\item{value_var}{An abundance value to use with \code{hclust}}
}
\value{
A \code{triple_omic} with clustered features or samples.
}
\description{
Sort a \code{triple_omic} object using hierarchical clustering
}
