% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_sort.R
\name{sortInput}
\alias{sortInput}
\title{Sort Input}
\usage{
sortInput(id, sort_table)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{sort_table}{table to sort}
}
\value{
A \code{shiny} UI
}
\description{
UI components for the sort module.
}
