% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_sort.R
\name{sortServer}
\alias{sortServer}
\title{Sort Server}
\usage{
sortServer(id, tomic, sort_table, valid_sort_vars = NULL, value_var = NULL)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}

\item{sort_table}{samples or features}

\item{valid_sort_vars}{variables available for categorical arranging}

\item{value_var}{An abundance value to use with \code{hclust}}
}
\value{
A sorted \code{tomic} object.
}
\description{
Server components for the sort module.
}
