% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_classes.R
\name{tidy_to_triple}
\alias{tidy_to_triple}
\title{Tidy omic to triple omic}
\usage{
tidy_to_triple(tidy_omic)
}
\arguments{
\item{tidy_omic}{an object of class tidy_omic produced by
\code{\link{create_tidy_omic}}}
}
\description{
Tidy omic to triple omic
}
\examples{
tidy_to_triple(brauer_2008_tidy)
}
