% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutates.R
\name{tomic_sort_status}
\alias{tomic_sort_status}
\title{T* Omic Sort Status}
\usage{
tomic_sort_status(tomic)
}
\arguments{
\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}
}
\value{
length 1 character string indicating whether the \code{tomic}
  is sorted.
}
\description{
Determine whether features &/or samples have been sorted and stored as
  ordered_featureId and ordered_sampleId.
}
\examples{

tomic_sort_status(brauer_2008_tidy)
}
