% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_classes.R
\name{tomic_to}
\alias{tomic_to}
\title{T* Omic To}
\usage{
tomic_to(tomic, to_class)
}
\arguments{
\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}

\item{to_class}{The class to return, either \code{tidy_omic} or
\code{triple_omic}}
}
\value{
tomic transformed to \code{to_class} class (or un-transformed if
  it started that way).
}
\description{
Takes in any \code{romic} reprsentation of a dataset and returns a specific
  representation.
}
\examples{

tomic_to(brauer_2008_tidy, "triple_omic")
}
