% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ggbiv.R
\name{try_brushedPoints}
\alias{try_brushedPoints}
\title{Try brushedPoints}
\usage{
try_brushedPoints(...)
}
\arguments{
\item{...}{args to pass to \link[shiny]{brushedPoints}}
}
\value{
a df of brushed points
}
\description{
This function wraps brushedPoints in a try statement to catch cases where
  the brushing is out-of-sync with the df that is selected.
}
