% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutates.R
\name{update_sample_factors}
\alias{update_sample_factors}
\title{Update Sample Factors}
\usage{
update_sample_factors(tomic, factor_levels)
}
\arguments{
\item{tomic}{Either a \code{tidy_omic} or \code{triple_omic} object}

\item{factor_levels}{a character vector specifying the ordering of factor levels.}
}
\value{
a tomic object with updated sample metadata
}
\description{
Update sample metadata to order categorical variables based on a
specified factor order.
}
\examples{
update_sample_factors(
  brauer_2008_tidy, list(nutrient = c("G", "N", "P", "S", "L", "U"))
)

}
