% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_partial_match}
\alias{var_partial_match}
\title{Var Partial Match}
\usage{
var_partial_match(x, df)
}
\arguments{
\item{x}{a variable name or regex match to a variable name}

\item{df}{a data.frame or tibble}
}
\value{
a single variable from df
}
\description{
Partial string matching of a provided variable to the variables available
in a table
}
