% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ggplot.R
\name{ggplotOutput}
\alias{ggplotOutput}
\title{ggplot Output}
\usage{
ggplotOutput(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
A \code{shiny} UI
}
\description{
UI components for the ggplot module.
}
