% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plotsaver.R
\name{plotsaverServer}
\alias{plotsaverServer}
\title{Plot Saver Server}
\usage{
plotsaverServer(id, grob)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{grob}{a ggplot2 plot}
}
\value{
None
}
\description{
Server components for the plot saver module.
}
