% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romic.R
\docType{package}
\name{romic}
\alias{romic}
\title{\code{romic} package}
\description{
'romic' represents high-dimensional data as tables of features,
  samples and measurements, and a design list for tracking the meaning of
  individual variables. Using this format, filtering, normalization, and
  other transformations of a dataset can be carried out in a flexible
  manner. 'romic' takes advantage of these transformations to create
  interactive shiny apps for exploratory data analysis such as an
  interactive heatmap.
}
