% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plotsaver.R
\name{shiny_plotsaver_test}
\alias{shiny_plotsaver_test}
\title{Shiny Plot Saver Test}
\usage{
shiny_plotsaver_test()
}
\value{
a \code{shiny} app
}
\description{
Test the shiny plotsaver module as a stand-alone application.
}
\examples{

if (interactive()) {
  shiny_plotsaver_test()
}
}
