% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_classes.R
\name{tidy_to_triple}
\alias{tidy_to_triple}
\title{Tidy omic to triple omic}
\usage{
tidy_to_triple(tidy_omic)
}
\arguments{
\item{tidy_omic}{an object of class tidy_omic produced by
\code{\link{create_tidy_omic}}}
}
\value{
A \code{triple_omic} object as created by
  \code{\link{create_triple_omic}}
}
\description{
Convert a \code{tidy_omic} object into a \code{triple_omic} object.
}
\details{
The \code{data} table will be converted into \code{features},
  \code{samples}, and \code{measurements} tables using the \code{design}
  to determine which variables belong in each table. The \code{design}
  will be preserved as-is.
}
\examples{
tidy_to_triple(brauer_2008_tidy)
}
